Rename https to web
